requirements.txt
elasticsearch==7.0.3
Flask==1.0
flask-bootstrap==3.3.7.1
flask-login
flask-mail==0.9.1
flask-migrate
flask-moment==0.7.0
flask-sqlalchemy
flask-wtf
pyjwt==1.7.1
python-dotenv